Back/django-battleships-master/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
asgiref==3.8.1

# Вспомогательные библиотеки (если ты используешь линтеры/тесты)
flake8==7.2.0
pyflakes==3.3.2
pycodestyle==2.13.0
mccabe==0.7.0
mock==5.2.0

# Зависимости других пакетов (опционально)
setuptools==65.5.0
pip==22.3.1
typing-extensions==4.13.2
